backend/requirements.txt
flask
flask_cors
fastapi
python-socketio[asgi_client]
# uvicorn
eventlet
flask-socketio==5.3.6
# python-socketio==5.11.1
python-engineio==4.9.1
uvicorn==0.29.0